refactor(step-3):피드백 반영하여 패키지 구조 변경
- 기존 domain에 다 들어있던 구조에서 controller, model, repository, service로 구분
- service 내에서는 impl 패키지를 만들어서 분리
