+++ src/main/java/kr/codesqaud/cafe/controller/UserController.java
@@ -1,6 +1,5 @@
 package kr.codesqaud.cafe.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,8 +8,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import kr.codesqaud.cafe.domain.JoinService;
-import kr.codesqaud.cafe.user.User;
+import kr.codesqaud.cafe.service.JoinService;
+import kr.codesqaud.cafe.model.User;
 
 @Controller
 public class UserController {
