+++ src/main/java/kr/codesqaud/cafe/model/User.java
@@ -1,6 +1,4 @@
-package kr.codesqaud.cafe.user;
-
-import java.util.Objects;
+package kr.codesqaud.cafe.model;
 
 public class User {
     private String id;
@@ -25,19 +23,9 @@ public class User {
         this.index = index;
     }
 
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public void setName(String name) {
+    public void updateUserInfo(String name, String password, String email) {
         this.name = name;
-    }
-
-    public void setPassword(String password) {
         this.password = password;
-    }
-
-    public void setEmail(String email) {
         this.email = email;
     }
