+++ src/main/java/kr/codesqaud/cafe/repository/ArticleRepository.java
@@ -1,9 +1,9 @@
-package kr.codesqaud.cafe.domain;
+package kr.codesqaud.cafe.repository;
 
 import java.util.List;
 import java.util.Optional;
 
-import kr.codesqaud.cafe.user.Article;
+import kr.codesqaud.cafe.model.Article;
 
 public interface ArticleRepository {
     public void addArticle(Article article);
