+++ src/main/java/kr/codesqaud/cafe/repository/MemoryArticleRepository.java
@@ -1,10 +1,10 @@
-package kr.codesqaud.cafe.domain;
+package kr.codesqaud.cafe.repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import kr.codesqaud.cafe.user.Article;
+import kr.codesqaud.cafe.model.Article;
 
 public class MemoryArticleRepository implements ArticleRepository {
     private List<Article> articleList = new ArrayList<>();
