+++ src/main/java/kr/codesqaud/cafe/repository/MemoryUserRepository.java
@@ -1,10 +1,10 @@
-package kr.codesqaud.cafe.domain;
+package kr.codesqaud.cafe.repository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-import kr.codesqaud.cafe.user.User;
+import kr.codesqaud.cafe.model.User;
 
 public class MemoryUserRepository implements UserRepository {
     private ArrayList<User> userList = new ArrayList<>();
@@ -33,8 +33,6 @@ public class MemoryUserRepository implements UserRepository {
         //비밀번호 일치 확인
         user.validate(password);
 
-        user.setPassword(newPassword);
-        user.setEmail(email);
-        user.setName(name);
+        user.updateUserInfo(name, newPassword, email);
     }
 }
